Загрузка данных отдельно, изменения данных (без api)

+++ api/KpoApi/KpoApi.Domain/Entities/Organizations.cs
@@ -10,5 +10,5 @@ public record Organization
     public int SsmpNumber { get; init; }
     public string SsmpAdress { get; init; }
     public string PhoneContactName { get; init; }
-    public int PhoneNumber { get; init; }
+    public long PhoneNumber { get; init; }
 }
